Elasticsearch indexing and searching changes, requirement.txt changes, Data Content Labelling
