+++ app/templates/index.html
@@ -1,3 +1,8 @@
+<!--
+Authors:
+Supritha Amudhu
+-->
+
 <!doctype html>
 <html lang="en">
 <head>
