+++ requirement.txt
@@ -41,6 +41,7 @@ nbformat==4.4.0
 notebook==5.7.0
 pandocfilters==1.4.2
 parso==0.3.1
+Pattern==2.6
 pattern3==3.0.0
 pdfminer.six==20181108
 pdfminer3k==1.3.1
@@ -78,3 +79,11 @@ webencodings==0.5.1
 Werkzeug==0.14.1
 widgetsnbextension==3.4.2
 zc.lockfile==1.4
+matplotlib==2.2.2
+numpy==1.13.3
+selenium==3.14.1
+pandas==0.20.3
+gensim==3.4.0
+nltk==3.2.5
+scikit_learn==0.20.2
+
